pyproject.toml
[project]
name = "100-days-of-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorgram-py>=1.2.0",
    "pandas>=2.2.3",
    "prettytable>=3.15.1",
    "pyperclip>=1.9.0",
]
